+++ src/commands/build-react-component.js
@@ -12,6 +12,7 @@ export default function buildModule(args, cb) {
   let config = {
     babel: {
       presets: [require.resolve('babel-preset-react')],
+      stage: 1,
     }
   }